Persist tasks using H2 and move greeting classes
