+++ src/main/java/nl/cge/greeting/GreetingResponse.java
@@ -1,3 +1,3 @@
-package nl.cge;
+package nl.cge.greeting;
 
 public record GreetingResponse(String message) {}
