+++ src/main/java/nl/cge/tasks/TaskService.java
@@ -1,21 +1,24 @@
 package nl.cge.tasks;
 
 import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Inject;
+import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 @ApplicationScoped
 public class TaskService {
-    private final List<Task> tasks = new ArrayList<>();
+    @Inject
+    EntityManager em;
 
-    public synchronized void addTask(String name, String description, LocalDate endDate) {
-        tasks.add(new Task(name, description, endDate));
+    @Transactional
+    public void addTask(String name, String description, LocalDate endDate) {
+        em.persist(new Task(name, description, endDate));
     }
 
     public List<Task> listTasks() {
-        return Collections.unmodifiableList(tasks);
+        return em.createQuery("from Task", Task.class).getResultList();
     }
 }
