+++ src/main/resources/application.properties
@@ -1,0 +1,3 @@
+quarkus.datasource.db-kind=h2
+quarkus.datasource.jdbc.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1
+quarkus.hibernate-orm.database.generation=drop-and-create
